refactored powergrid column css classes
there is now a `cssClasses()` method that provides mixins a handle to
insert new css classes for both the table cells and headers.

the method that used to be `cssClass()` is now `columnClass()`, which
basically just returns something like `col-<column name>`.

this also adds a `.first` and `.last` class to the first and last
columns.