View uses .render *nearly* always during init
we want the View class to always call .render during init with the one
exception where the view is being instantiated around an existing DOM
structure.

the one exception is actually an anti-pattern for views, and if it's the
case, it should be using a Widget, for example if we wanted to
instantiate a ToggleGroup around existing DOM elements, this is a valid
use case, however it should instantiate a Widget, not a View.