prevent infinite loop in powergridpager
this loop wouldn't occur in normal situations, but if a collection's
'total' parameter is undefined, then it has the potential to set the
powergridpager's 'pages' option to 'NaN', which is !== to itself, which
triggers a change, which set's the 'pages' to NaN again... etc.

we could protect against this in a broad sense by checking for NaN in
asSettable, and handling it the 'expected' way (if something is NaN, and
the user sets it to NaN, nothing happens). this would help find these
sorts of bugs, but in most cases would just slow things down, so i'm
hesitant.

pertains to siq/mesh#19